+++ OkHttpFinal/src/main/java/cn/finalteam/okhttpfinal/FileDownloadTask.java
@@ -44,7 +44,7 @@ public class FileDownloadTask extends AsyncTask<Void, Long, Boolean> {
 
     public FileDownloadTask(String url, File target, FileDownloadCallback callback) {
         this.url = url;
-        this.okHttpClient = OkHttpFinal.getInstance().getOkHttpClient();
+        this.okHttpClient = OkHttpFinal.getInstance().getOkHttpClientBuilder().build();
         this.callback = callback;
         this.target = target;